+++ source/isaaclab_assets/isaaclab_assets/robots/kuka_allegro.py
@@ -16,6 +16,9 @@ Reference:
 
 """
 
+from isaaclab_newton.sim.schemas import NewtonArticulationCfg
+from isaaclab_physx.sim.schemas import PhysxArticulationCfg
+
 import isaaclab.sim as sim_utils
 from isaaclab.actuators import ImplicitActuatorCfg
 from isaaclab.assets.articulation import ArticulationCfg
@@ -37,13 +40,16 @@ KUKA_ALLEGRO_CFG = ArticulationCfg(
             max_angular_velocity=1000.0,
             max_depenetration_velocity=1000.0,
         ),
-        articulation_props=sim_utils.ArticulationRootPropertiesCfg(
-            enabled_self_collisions=True,
-            solver_position_iteration_count=32,
-            solver_velocity_iteration_count=1,
-            sleep_threshold=0.005,
-            stabilization_threshold=0.0005,
-        ),
+        articulation_props=[
+            PhysxArticulationCfg(
+                enabled_self_collisions=True,
+                solver_position_iteration_count=32,
+                solver_velocity_iteration_count=1,
+                sleep_threshold=0.005,
+                stabilization_threshold=0.0005,
+            ),
+            NewtonArticulationCfg(self_collision_enabled=True),
+        ],
         joint_drive_props=sim_utils.JointDrivePropertiesCfg(drive_type="force"),
     ),
     init_state=ArticulationCfg.InitialStateCfg(